requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
click==8.0.1
Flask==2.0.1
geffnet==1.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
Pillow==8.3.1
torch==1.9.0+cpu
torchvision==0.10.0+cpu
typing-extensions==3.10.0.0
Werkzeug==2.0.1
